requirements.txt
alembic==1.4.1
Boruta==0.3
certifi==2021.5.30
charset-normalizer==2.0.5
click==8.0.1
cloudpickle==2.0.0
cycler==0.10.0
Cython==0.29.24
databricks-cli==0.15.0
dataclasses==0.8
docker==5.0.2
entrypoints==0.3
et-xmlfile==1.1.0
Flask==2.0.1
gitdb==4.0.7
GitPython==3.1.18
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
imbalanced-learn==0.8.0
imblearn==0.0
importlib-metadata==4.8.1
ITMO-FS==0.3.3
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.6.3
Mako==1.1.5
MarkupSafe==2.0.1
matplotlib==3.3.4
mlflow==1.20.2
numpy==1.19.5
opencv-python==4.5.3.56
openpyxl==3.0.8
packaging==21.0
pandas==1.1.5
patsy==0.5.1
Pillow==8.3.2
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.2
protobuf==3.18.0
pymrmre==1.0.7
pyparsing==2.4.7
python-dateutil==2.8.2
python-docx==0.8.11
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
qpsolvers==1.6.1
quadprog==0.1.8
querystring-parser==1.2.4
requests==2.26.0
scikit-learn==0.24.2
scikit-posthocs==0.6.7
scipy==1.5.4
seaborn==0.11.2
six==1.16.0
smmap==4.0.0
SQLAlchemy==1.4.23
sqlparse==0.4.2
statsmodels==0.12.2
tabulate==0.8.9
threadpoolctl==2.2.0
typing-extensions==3.10.0.2
urllib3==1.26.6
websocket-client==1.2.1
Werkzeug==2.0.1
xgboost==1.4.2
zipp==3.5.0
